cleaning for pull-request, which include a new activation named GHH, which was used for 2 recents publications and generalize relu / prelu / maxout
