+++ .gitignore
@@ -161,10 +161,10 @@ pip-log.txt
 #Mr Developer
 .mr.developer.cfg
 
+
 # Mac crap
 .DS_Store
-
-build
+build #usual folder made for cmake...
 
 
 #################
