+++ example/build.bat
@@ -1,2 +1,2 @@
-g++ -std=c++11 -Wall -Werror -Wextra -pedantic -I ../tiny_cnn -O3 train.cpp -o train
-g++ -std=c++11 -Wall -Werror -Wextra -pedantic -I ../tiny_cnn -O3 test.cpp -o test
+g++ -std=c++11 -Wall -Werror -Wextra -pedantic -I ../tiny_cnn -O3 sample_train.cpp -o sample_train
+g++ -std=c++11 -Wall -Werror -Wextra -pedantic -I ../tiny_cnn -O3 sample_test.cpp -o sample_test
